kafka-2/backend/requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
kafka-python==2.2.12
mysqlclient==2.1.1
redis==6.2.0
sqlparse==0.5.3
typing_extensions==4.14.0
